sync_music: Remove faulty optimization (using sort)
The initial version sorts the list of input files by action so that
fast running actions (skip/copy) are processed before slow running
actions so that the majority of items can be processed quickly.
The sort is broken as there is no total order for actions defined.
In addition when executing on multiple processors, the items are
distributed in nproc blocks and so the sorting becomes useless.