sync_music: Create destination directory
Create the destination directory when reading the settings. This
improves usability as the user doesn't have to enter a destination
directory that fully exists.

If there are no input files or if program aborts due to any reason
the created directory stays empty on the disk and is not deleted.