hack: Set album to 'Various Artists' if albumartist is empty
For the albumartist artist hack, when the albumartist field is empty
the artist is not overridden and thus players show this artist in their
artist view.

The purpose of this hack is however to show only album artists in the
artist view. Therefore override the artist field with 'Various Artists'
for such files.